+++ frontend/pages/my-collections.tsx
@@ -2,13 +2,13 @@ import Image from 'next/image';
 import withTransition from '../components/withTransition';
 import { ConnectToWalletBanner } from '../components/ConnectToWalletBanner';
 import { Loader, Banner } from '../components';
-import { addressToCreatorImage, shortenAddress } from '../utils';
+import { addressToCreatorImage, shortenAddress } from '../services/utils';
 import { useAlephiumConnectContext } from '@alephium/web3-react';
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import Link from 'next/link';
 import { motion } from 'framer-motion';
-import { shortenName } from '../utils';
+import { shortenName } from '../services/utils';
 import { backendUrl } from '../../configs/nft';
 import { NFTCollectionMetadata, fetchNFTCollectionMetadata } from '../../shared/nft-collection';
 
@@ -132,7 +132,7 @@ const MyCollections = () => {
         <div className="sm:px-4 p-12 w-full ml-32 mr-32 md:ml-20 md:mr-20 sm:ml-10 sm:mr-10 mt-3">
           <div className="grid-container">
             {collections.map((collection) => (
-              <NFTCollectionCard key={collection._id} collection={collection}/>
+              <NFTCollectionCard key={collection._id} collection={collection} />
             ))}
           </div>
         </div>
